requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==42.0.5
bcrypt==5.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
jinja2==3.1.4
python-dotenv==1.0.1
email-validator==2.1.0.post1